heapsafe: Only declare mem* if they are present
Newer versions of compilers (e.g. clang and gcc) declare some standard library
functions as macros. If these mem* prototypes are given, the compiler applies
the macro and then subsequently produces a lot of strange syntax errors.

The quick solution used here is to only declare the prototypes if they are not
macros. This is not the best solution. Rather, we would like to somehow use the
warnings of hs_untyped even if the macros are used. But this will need some
thought.